backend/requirements.txt
pydantic_settings
python-dotenv
fastapi
uvicorn
rich
